pyproject.toml
[project]
name = "rlm-legal-docs"
version = "0.1.0"
description = "Legal document query tool using Recursive Language Models"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Rohan George", email = "rohankgeorge@users.noreply.github.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Legal Industry",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "rlms>=0.1.0",
    "customtkinter>=5.2.2",
    "python-docx>=1.2.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "langextract>=1.1.0",
]

[project.urls]
Repository = "https://github.com/rohankgeorge/RLM-Legal-Implementation"
"Original RLM Project" = "https://github.com/alexzhang13/rlm"

[project.scripts]
rlm-legal-gui = "rlm_legal_docs.cli:run_gui"
rlm-legal-query = "rlm_legal_docs.cli:run_query"

[project.optional-dependencies]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
    "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rlm_legal_docs", "rlm_legal_docs.*"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
]
test = [
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
